+++ analytics/analytics.go
@@ -19,7 +19,7 @@ import (
 	"github.com/earthly/earthly/cliutil"
 	"github.com/earthly/earthly/fileutil"
 	"github.com/earthly/earthly/gitutil"
-	"github.com/earthly/earthly/syncutil"
+	"github.com/earthly/earthly/util/syncutil"
 
 	"github.com/google/uuid"
 	"github.com/pkg/errors"
