+++ buildcontext/git.go
@@ -15,7 +15,7 @@ import (
 	"github.com/earthly/earthly/llbutil"
 	"github.com/earthly/earthly/llbutil/pllb"
 	"github.com/earthly/earthly/stringutil"
-	"github.com/earthly/earthly/syncutil/synccache"
+	"github.com/earthly/earthly/util/syncutil/synccache"
 
 	"github.com/moby/buildkit/client/llb"
 	gwclient "github.com/moby/buildkit/frontend/gateway/client"
