+++ buildcontext/local.go
@@ -9,7 +9,7 @@ import (
 	"github.com/earthly/earthly/gitutil"
 	"github.com/earthly/earthly/llbutil"
 	"github.com/earthly/earthly/llbutil/pllb"
-	"github.com/earthly/earthly/syncutil/synccache"
+	"github.com/earthly/earthly/util/syncutil/synccache"
 	"github.com/moby/buildkit/client/llb"
 	"github.com/pkg/errors"
 )
