+++ buildcontext/resolver.go
@@ -11,7 +11,7 @@ import (
 	"github.com/earthly/earthly/domain"
 	"github.com/earthly/earthly/gitutil"
 	"github.com/earthly/earthly/llbutil/pllb"
-	"github.com/earthly/earthly/syncutil/synccache"
+	"github.com/earthly/earthly/util/syncutil/synccache"
 	gwclient "github.com/moby/buildkit/frontend/gateway/client"
 	"github.com/pkg/errors"
 )
